Locacao mais ou menos funcionando.

+++ LocadoraVeiculo/LocadoraDeVeiculos.Infra.ORM/GrupoVeiculoModule/GrupoVeiculoConfigurations.cs
@@ -16,23 +16,19 @@ namespace LocadoraDeVeiculos.Infra.Configurations.GrupoVeiculoModule
            
             builder.ToTable("TBTiposVeiculo");
 
-            builder.HasKey(grupoveiculo => grupoveiculo.Id);
-            
+            builder.HasKey(grupoveiculo => grupoveiculo.Id);            
 
             builder.Property(grupoveiculo => grupoveiculo.NomeTipo).HasColumnType("VARCHAR(64)").IsRequired();
 
-            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPControlado).HasColumnType("DECIMAL").IsRequired();
+            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPControlado).HasColumnType("FLOAT").IsRequired();
 
-            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPDiario).HasColumnType("DECIMAL").IsRequired();
+            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPDiario).HasColumnType("FLOAT").IsRequired();
 
-            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPLivre).HasColumnType("DECIMAL").IsRequired();
+            builder.Property(grupoveiculo => grupoveiculo.ValorDiarioPLivre).HasColumnType("FLOAT").IsRequired();
 
-            builder.Property(grupoveiculo => grupoveiculo.ValorKmRodadoPControlado).HasColumnType("DECIMAL").IsRequired();
+            builder.Property(grupoveiculo => grupoveiculo.ValorKmRodadoPControlado).HasColumnType("FLOAT").IsRequired();
 
-            builder.Property(grupoveiculo => grupoveiculo.ValorKmRodadoPDiario).HasColumnType("DECIMAL").IsRequired();
-
-
-
+            builder.Property(grupoveiculo => grupoveiculo.ValorKmRodadoPDiario).HasColumnType("FLOAT").IsRequired();
 
         }
     }
